Update Interface for modified Chebyshev moment-based quadrature
For the two strings `clenshawcurtis` and `fejer`, there are now three methods each:

`plan_*`, `*nodes`, `*weights`, where the Fejer methods are followed by a `1` or `2` to indicate the type.

This is a better interface because `*nodes` is more extensible: different Chebyshev moments may be added for new quadrature rules. In particular, this commit also adds `log` moments.